Fix issues in IDEEP conv ops and etc.
