+++ caffe2/ideep/utils/ideep_operator.h
@@ -1,25 +1,8 @@
-/**
- * Copyright (c) 2016-present, Facebook, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
 #pragma once
 
 #include <ideep.hpp>
 #include <caffe2/core/operator.h>
 #include <caffe2/proto/caffe2.pb.h>
-#include <caffe2/operators/conv_pool_op_base.h>
 
 namespace caffe2 {
 
@@ -113,93 +96,12 @@ class IDEEPOperator : public OperatorBase {
   StorageOrder order_;
 };
 
-
-class IDEEPConvPoolOpBase : public ConvPoolOpBase<IDEEPContext> {
- public:
-  IDEEPConvPoolOpBase(const OperatorDef& operator_def, Workspace* ws)
-     : ConvPoolOpBase<IDEEPContext>(operator_def, ws) {
-    OPERATOR_NEEDS_FEATURE(
-        order_ == StorageOrder::NCHW, "Unsupported storage order.");
-  }
-  virtual ~IDEEPConvPoolOpBase() {}
-
-  inline const ideep::tensor& Input(int index) {
-    return OperatorBase::template Input<ideep::tensor>(index);
-  }
-  inline ideep::tensor* Output(int index) {
-    return OperatorBase::template Output<ideep::tensor>(index);
-  }
-
-  ideep::tensor::dims pad_tl() {
-    return {pad_t(), pad_l()};
-  }
-
-  ideep::tensor::dims pad_br() {
-    return {pad_b(), pad_r()};
-  }
-
-  ideep::tensor::dims CalcOutputDims(
-      const ideep::tensor& input,
-      int output_channel) {
-    CAFFE_ENFORCE(input.get_descriptor().get_size() > 0);
-
-    bool channel_first;
-    int N = input.get_dim(0);
-    ideep::tensor::dims output_dims;
-
-    auto input_dims = input.get_dims();
-    vector<TIndex> input_Tdims (input_dims.begin(), input_dims.end());
-    InferOutputSize(
-        input_Tdims,
-        output_channel,
-        order_,
-        global_pooling_,
-        legacy_pad_,
-        N,
-        kernel_,
-        output_dims,
-        dilation_,
-        stride_,
-        pads_,
-        channel_first);
-
-    if (channel_first) {
-      output_dims.insert(output_dims.begin(), {N, output_channel});
-    } else {
-      output_dims.insert(output_dims.begin(), N);
-      output_dims.push_back(output_channel);
-    }
-
-    return output_dims;
-  }
-
-  bool RunOnDevice() override {
-    if (!global_pooling_) {
-      for (int dim = 0; dim < kernel_.size(); ++dim) {
-        CAFFE_ENFORCE_GT(kernel_[dim], 0);
-      }
-    }
-
-    try {
-      return RunOnDeviceWithOrderNCHW();
-    } catch (ideep::error& e) {
-      VLOG(1) << "IDEEP error:" << e.message; 
-      throw;
-    }
-  }
-};
-
 #define USE_IDEEP_OPERATOR_FUNCTIONS()                                         \
   USE_OPERATOR_BASE_FUNCTIONS;                                                 \
   /* using override */ using IDEEPOperator::Input;                             \
   /* using override */ using IDEEPOperator::Output;                            \
   /* using override */ using IDEEPOperator::order_;                            \
   /* using override */ using IDEEPOperator::context_;
-
-#define USE_IDEEP_CONV_POOL_BASE_FUNCTIONS()                                   \
-  USE_OPERATOR_BASE_FUNCTIONS;                                                 \
-  /* using override */ using IDEEPConvPoolOpBase::Input;                       \
-  /* using override */ using IDEEPConvPoolOpBase::Output;
 
 #define USE_SIMPLE_IDEEP_CTOR_DTOR(name)                                       \
   name(const OperatorDef& operator_def, Workspace* ws)                         \
